organize files around features
