+++ usersAuth/auth.controllers.js
@@ -1,4 +1,4 @@
-const User = require('../models/user.model');
+const User = require('./user.model');
 const jwt = require('jsonwebtoken')
 
 const createUser = async (req, res)=>{
